# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Önder Grup Back End API",
    "version": "2.0.0",
    "description": "Önder Grup Back End API for Embedded & Hydraulic",
    "contact": {
      "name": "Halil İbrahim Direktör",
      "email": "hidirektor@gmail.com",
      "url": "https://hidirektor.com.tr"
    }
  },
  "servers": [
    {
      "url": "http://ondergrup.hidirektor.com.tr:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/getProfilePhoto/{userName}": {
      "get": {
        "summary": "Retrieves the URL of a user's profile photo by userName.",
        "tags": [
          "User Profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userName",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The username to find the profile photo URL."
          }
        ],
        "responses": {
          "200": {
            "description": "Profile photo retrieved successfully.",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Profile photo not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  },
...
}